+++ auth/presentation/src/test/kotlin/me/proton/core/auth/presentation/viewmodel/signup/SignupViewModelTest.kt
@@ -119,7 +119,6 @@ class SignupViewModelTest : ArchTest, CoroutinesTest {
             keyStoreCrypto,
             plansOrchestrator,
             paymentsOrchestrator,
-            clientIdProvider,
             humanVerificationManager,
             performLogin,
             challengeManager,